Moves footer components to ui folder
